refactor: 코드 전반 리팩터링
Text 컴포넌트 통일, util 및 상수 분리, Spacing 태그 톻일
