+++ frontend/src/api/post.ts
@@ -1,5 +1,7 @@
 import { api } from '.';
+import { CouponSettingDto } from '../pages/Admin/CustomCouponDesign';
 import { StampFormData } from '../pages/Admin/EarnStamp';
+import { CafeFormData } from '../pages/Admin/RegisterCafe';
 
 export const postEarnStamp = async ({
   earningStampCount,
@@ -20,4 +22,23 @@ export const postIssueCoupon = async (customerId: string) => {
   return await api
     .post(`/customers/${customerId}/coupons`, { cafeId: '1' })
     .then((response) => response.json());
+};
+
+export const postCouponSetting = async (couponConfig: CouponSettingDto) => {
+  return await api.post('/coupon-setting', couponConfig);
+};
+
+export const postRegisterCafe = async ({
+  ownerId,
+  businessRegistrationNumber,
+  name,
+  roadAddress,
+  detailAddress,
+}: CafeFormData) => {
+  await api.post(`/cafes/${ownerId}`, {
+    businessRegistrationNumber,
+    name,
+    roadAddress,
+    detailAddress,
+  });
 };
