+++ frontend/src/pages/Admin/CustomCouponDesign/ChoiceTemplate/style.tsx
