+++ frontend/src/pages/Admin/CustomCouponDesign/CustomStampSection/style.tsx
