+++ frontend/src/pages/Admin/ModifyCouponPolicy/RadioInput/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { RadioInputsContainer, Label, LabelText, RadioInput } from './RadioInputs.style';
-import { Spacing } from '../../../style/layout/common';
+import { RadioInputsContainer, Label, LabelText, RadioInput } from './style';
+import { Spacing } from '../../../../style/layout/common';
 
 interface RadioInputsProps {
   setValue: React.Dispatch<React.SetStateAction<string>>;
