+++ frontend/src/style/layout/common.tsx
@@ -17,22 +17,3 @@ export const RowSpacing = styled.div<StyledSpacingProps>`
   width: ${({ $size }) => `${$size}px`};
   height: 100%;
 `;
-export const Title = styled.h2`
-  font-size: 36px;
-  font-weight: 700;
-`;
-
-export const SectionTitle = styled.p`
-  font-size: 16px;
-  font-weight: 700;
-`;
-
-export const SubTitle = styled.h3`
-  font-size: 24px;
-  font-weight: 700;
-`;
-
-export const Text = styled.span`
-  font-size: 12px;
-  color: ${({ theme }) => theme.colors.black};
-`;
